% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuCellAI_mouse.R, R/hello.R
\name{compensation}
\alias{compensation}
\title{Title}
\usage{
compensation(raw_score, compensation_matrix)

compensation(raw_score, compensation_matrix)
}
\arguments{
\item{compensation_matrix}{}
}
\value{



}
\description{
Title

Title
}
