% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuCellAI_mouse.R, R/hello.R
\name{gene_ref_weight_mt}
\alias{gene_ref_weight_mt}
\title{Title}
\usage{
gene_ref_weight_mt(marker_ls, marker_reccur)

gene_ref_weight_mt(marker_ls, marker_reccur)
}
\arguments{
\item{marker_reccur}{}
}
\value{



}
\description{
Title

Title
}
