% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getResult}
\alias{getResult}
\title{Title}
\usage{
getResult(
  sample,
  data_type,
  marker_exp,
  paper_marker,
  ref_pre,
  cell_cor_new_mat,
  compensation_matrix
)
}
\arguments{
\item{compensation_matrix}{}
}
\value{

}
\description{
Title
}
