% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuCellAI_mouse.R, R/hello.R
\name{immune_infiltate_calculate}
\alias{immune_infiltate_calculate}
\title{Title}
\usage{
immune_infiltate_calculate(exp, gene_name, data_type, pre_result)

immune_infiltate_calculate(exp, gene_name, data_type, pre_result)
}
\arguments{
\item{pre_result}{}
}
\value{



}
\description{
Title

Title
}
