% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuCellAI_mouse.R, R/hello.R
\name{layer_norm}
\alias{layer_norm}
\title{Title}
\usage{
layer_norm(pre_result, cell_ls)

layer_norm(pre_result, cell_ls)
}
\arguments{
\item{cell_ls}{}
}
\value{



}
\description{
Title

Title
}
