% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuCellAI_mouse.R, R/hello.R
\name{marker_times_count}
\alias{marker_times_count}
\title{Title}
\usage{
marker_times_count(paper_marker)

marker_times_count(paper_marker)
}
\arguments{
\item{paper_marker}{}
}
\value{



}
\description{
Title

Title
}
