% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuCellAI_mouse.R, R/hello.R
\name{sample_ratio}
\alias{sample_ratio}
\title{Title}
\usage{
sample_ratio(
  data,
  marker_exp,
  marker_tag_mat,
  data_type,
  gene_weight_mt,
  ref_pre,
  cell_cor_new_mat
)

sample_ratio(
  data,
  marker_exp,
  marker_tag_mat,
  data_type,
  gene_weight_mt,
  ref_pre,
  cell_cor_new_mat
)
}
\arguments{
\item{cell_cor_new_mat}{}
}
\value{



}
\description{
Title

Title
}
