% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuCellAI_mouse.R, R/hello.R
\name{ImmuCellAI_mouse}
\alias{ImmuCellAI_mouse}
\title{Title}
\usage{
ImmuCellAI_mouse(
  sample,
  data_type,
  group_tag,
  customer,
  sig_file = NULL,
  exp_file = NULL
)

ImmuCellAI_mouse(
  sample,
  data_type,
  group_tag,
  customer,
  sig_file = NULL,
  exp_file = NULL
)
}
\arguments{
\item{exp_file}{}
}
\value{



}
\description{
Title

Title
}
